TASK 3
In this task, I created a Python program that generates strong and random passwords. The user enters the desired length, and the program builds a password using a mix of uppercase letters, lowercase letters, numbers, and special symbols. It also checks for invalid inputs and ensures better randomness by using the secrets library, making the passwords more secure and reliable for real use.