TASK 5
This project is a Python-based Contact Book application that helps users manage their contacts efficiently. It allows adding new contacts with details like name, phone number, email, and address, as well as viewing, searching, updating, and deleting existing contacts. All data is stored in a JSON file, ensuring that the information is saved even after the program is closed. The program also includes input validation for phone numbers and emails, along with a simple menu-driven interface that makes it beginner-friendly and easy to use.